ProcessExtensions.GetChildProcesses, GetDescendantProcesses, Kill(entireProcessTree), GetParentProcess

+++ src/Meziantou.Framework/ProcessExtensions.cs
@@ -2,42 +2,44 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
-using System.Threading;
-using System.Threading.Tasks;
+using Microsoft.Win32.SafeHandles;
 
 namespace Meziantou.Framework
 {
-    public static class ProcessExtensions
+    public static partial class ProcessExtensions
     {
-        private static bool IsWindows()
-        {
-#if NETSTANDARD2_0
-            return RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
-#elif NET461
-            return true;
-#else
-#error Platform not supported
-#endif
-        }
-
-        public static void KillProcessTree(this Process process)
+        // Mark as obsolete when merged https://github.com/dotnet/corefx/pull/31827
+        public static void Kill(this Process process, bool entireProcessTree = false)
         {
             if (process == null)
                 throw new ArgumentNullException(nameof(process));
 
+            if (!entireProcessTree)
+            {
+                process.Kill();
+                return;
+            }
+
             if (!IsWindows())
                 throw new PlatformNotSupportedException("Only supported on Windows");
 
-            var childProcesses = GetChildProcesses(process);
+            var childProcesses = GetDescendantProcesses(process);
+
+            if (!process.HasExited)
+            {
+                process.Kill();
+            }
+
             foreach (var childProcess in childProcesses)
             {
                 if (!childProcess.HasExited)
                 {
-                    childProcess.Kill();
+                    if (!childProcess.HasExited)
+                    {
+                        childProcess.Kill();
+                    }
                 }
             }
-
-            process.Kill();
 
             foreach (var childProcess in childProcesses)
             {
@@ -45,6 +47,19 @@ namespace Meziantou.Framework
             }
 
             process.WaitForExit();
+        }
+
+        public static IReadOnlyList<Process> GetDescendantProcesses(this Process process)
+        {
+            if (process == null)
+                throw new ArgumentNullException(nameof(process));
+
+            if (!IsWindows())
+                throw new PlatformNotSupportedException("Only supported on Windows");
+
+            var children = new List<Process>();
+            GetChildProcesses(process, children, int.MaxValue, 0);
+            return children;
         }
 
         public static IReadOnlyList<Process> GetChildProcesses(this Process process)
@@ -56,14 +71,33 @@ namespace Meziantou.Framework
                 throw new PlatformNotSupportedException("Only supported on Windows");
 
             var children = new List<Process>();
-            GetChildProcesses(process, children);
+            GetChildProcesses(process, children, 1, 0);
             return children;
         }
 
-        private static void GetChildProcesses(Process process, List<Process> children)
+        public static Process GetParentProcess(this Process process)
         {
-            var snapShotHandle = CreateToolhelp32Snapshot(SnapshotFlags.TH32CS_SNAPPROCESS, 0);
-            try
+            if (process == null)
+                throw new ArgumentNullException(nameof(process));
+
+            if (!IsWindows())
+                throw new PlatformNotSupportedException("Only supported on Windows");
+
+            var processId = process.Id;
+            foreach (var entry in GetProcesses())
+            {
+                if (entry.ProcessId == processId)
+                {
+                    return Process.GetProcessById(entry.ParentProcessId);
+                }
+            }
+
+            return null;
+        }
+
+        public static IEnumerable<ProcessEntry> GetProcesses()
+        {
+            using (var snapShotHandle = new SnapshotSafeHandle(CreateToolhelp32Snapshot(SnapshotFlags.TH32CS_SNAPPROCESS, 0), ownHandle: true))
             {
                 var entry = new ProcessEntry32
                 {
@@ -73,156 +107,67 @@ namespace Meziantou.Framework
                 var result = Process32First(snapShotHandle, ref entry);
                 while (result != 0)
                 {
-                    if (process.Id == entry.th32ParentProcessID)
-                    {
-                        try
-                        {
-                            var child = Process.GetProcessById((int)entry.th32ProcessID);
-                            children.Add(child);
-                            GetChildProcesses(child, children);
-                        }
-                        catch (ArgumentException)
-                        {
-                            // process might have exited since the snapshot, ignore it
-                        }
-                    }
-
+                    yield return new ProcessEntry(entry.th32ProcessID, entry.th32ParentProcessID);
                     result = Process32Next(snapShotHandle, ref entry);
                 }
             }
-            finally
-            {
-                if (snapShotHandle != IntPtr.Zero)
-                {
-                    CloseHandle(snapShotHandle);
-                }
-            }
         }
 
-        public static Task<ProcessResult> RunAsTask(string fileName, string arguments, CancellationToken cancellationToken = default)
+        private static void GetChildProcesses(Process process, List<Process> children, int maxDepth, int currentDepth)
         {
-            return RunAsTask(fileName, arguments, null, cancellationToken);
+            var entries = new List<ProcessEntry>(100);
+            foreach (var entry in GetProcesses())
+            {
+                entries.Add(entry);
+            }
+
+            GetChildProcesses(entries, process, children, maxDepth, currentDepth);
         }
 
-        public static Task<ProcessResult> RunAsTask(string fileName, string arguments, string workingDirectory, CancellationToken cancellationToken = default)
+        private static void GetChildProcesses(List<ProcessEntry> entries, Process process, List<Process> children, int maxDepth, int currentDepth)
         {
-            var psi = new ProcessStartInfo
+            var processId = process.Id;
+            foreach (var entry in entries)
             {
-                FileName = fileName,
-                Arguments = arguments,
-                WorkingDirectory = workingDirectory,
-                RedirectStandardError = true,
-                RedirectStandardOutput = true,
-                ErrorDialog = false,
-                UseShellExecute = false
-            };
-
-            return RunAsTask(psi, cancellationToken);
-        }
-
-        public static Task<ProcessResult> RunAsTask(this ProcessStartInfo psi, bool redirectOutput, CancellationToken cancellationToken = default)
-        {
-            if (redirectOutput)
-            {
-                psi.RedirectStandardError = true;
-                psi.RedirectStandardOutput = true;
-                psi.UseShellExecute = false;
-            }
-            else
-            {
-                psi.RedirectStandardError = false;
-                psi.RedirectStandardOutput = false;
-            }
-
-            return RunAsTask(psi, cancellationToken);
-        }
-
-        public static Task<ProcessResult> RunAsTask(this ProcessStartInfo psi, CancellationToken cancellationToken = default)
-        {
-            if (cancellationToken.IsCancellationRequested)
-                return Task.FromCanceled<ProcessResult>(cancellationToken);
-
-            var tcs = new TaskCompletionSource<ProcessResult>(TaskCreationOptions.RunContinuationsAsynchronously);
-            var logs = new List<ProcessOutput>();
-
-            var process = new Process();
-            process.StartInfo = psi;
-            process.EnableRaisingEvents = true;
-            process.Exited += (sender, e) =>
-            {
-                process.WaitForExit();
-                tcs.SetResult(new ProcessResult(process.ExitCode, logs));
-                process.Dispose();
-            };
-            process.ErrorDataReceived += (sender, e) =>
-            {
-                if (e.Data != null)
+                if (entry.ParentProcessId == processId)
                 {
-                    logs.Add(new ProcessOutput(ProcessOutputType.StandardError, e.Data));
-                }
-            };
-            process.OutputDataReceived += (sender, e) =>
-            {
-                if (e.Data != null)
-                {
-                    logs.Add(new ProcessOutput(ProcessOutputType.StandardOutput, e.Data));
-                }
-            };
-
-            if (!process.Start())
-                throw new InvalidOperationException($"Cannot start the process '{psi.FileName}'");
-
-            if (psi.RedirectStandardOutput)
-            {
-                process.BeginOutputReadLine();
-            }
-
-            if (psi.RedirectStandardError)
-            {
-                process.BeginErrorReadLine();
-            }
-
-            if (cancellationToken.CanBeCanceled)
-            {
-                cancellationToken.Register(() =>
-                {
-                    if (process.HasExited)
-                        return;
-
                     try
                     {
-                        if (IsWindows())
+                        var child = entry.ToProcess();
+                        children.Add(child);
+                        if (currentDepth < maxDepth)
                         {
-                            process.KillProcessTree();
-                        }
-                        else
-                        {
-                            process.Kill();
+                            GetChildProcesses(entries, child, children, maxDepth, currentDepth + 1);
                         }
                     }
-                    finally
+                    catch (ArgumentException)
                     {
-                        process.Dispose();
+                        // process might have exited since the snapshot, ignore it
                     }
-                });
+                }
             }
+        }
 
-            if (psi.RedirectStandardInput)
-            {
-                process.StandardInput.Close();
-            }
-
-            return tcs.Task;
+        private static bool IsWindows()
+        {
+#if NETSTANDARD2_0
+            return RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
+#elif NET461
+            return true;
+#else
+#error Platform not supported
+#endif
         }
 
         [DllImport("kernel32.dll", SetLastError = true)]
-        private static extern int CloseHandle(IntPtr hObject);
+        [return: MarshalAs(UnmanagedType.Bool)]
+        private static extern bool CloseHandle(IntPtr hObject);
 
         [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        private static extern int Process32First(IntPtr handle, ref ProcessEntry32 entry);
+        private static extern int Process32First(SnapshotSafeHandle handle, ref ProcessEntry32 entry);
 
         [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        private static extern int Process32Next(IntPtr handle, ref ProcessEntry32 entry);
+        private static extern int Process32Next(SnapshotSafeHandle handle, ref ProcessEntry32 entry);
 
         // https://msdn.microsoft.com/en-us/library/windows/desktop/ms682489.aspx
         [DllImport("kernel32.dll", SetLastError = true)]
@@ -232,16 +177,16 @@ namespace Meziantou.Framework
 
         // https://msdn.microsoft.com/en-us/library/windows/desktop/ms684839.aspx
         [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
-        private struct ProcessEntry32
+        internal struct ProcessEntry32
         {
 #pragma warning disable IDE1006 // Naming Styles
             public uint dwSize;
             public uint cntUsage;
-            public uint th32ProcessID;
+            public int th32ProcessID;
             public IntPtr th32DefaultHeapID;
             public uint th32ModuleID;
             public uint cntThreads;
-            public uint th32ParentProcessID;
+            public int th32ParentProcessID;
             public int pcPriClassBase;
             public uint dwFlags;
 
@@ -259,6 +204,20 @@ namespace Meziantou.Framework
             TH32CS_SNAPMODULE = 0x00000008,
             TH32CS_SNAPMODULE32 = 0x00000010,
             TH32CS_INHERIT = 0x80000000
+        }
+
+        private class SnapshotSafeHandle : SafeHandleZeroOrMinusOneIsInvalid
+        {
+            public SnapshotSafeHandle(IntPtr handle, bool ownHandle)
+                : base(ownHandle)
+            {
+                SetHandle(handle);
+            }
+
+            protected override bool ReleaseHandle()
+            {
+                return CloseHandle(handle);
+            }
         }
     }
 }
